remove OngoingHeader, move ReportHeader to headers folder
